Add restaurant component
